all: replace all occurrences of AddressBook with Concierge
The current application still uses the term "AddressBook" in many places.

However, this is misleading since this application is different from AddressBook.

Almost all relevant occurrences of "AddressBook" that were found with Intellij's search have been replaced with "Concierge". Functionality is yet to be tested. Some files still need to be renamed.